requirements.txt
hikari[speedups]
hikari-lightbulb
apscheduler
aiosqlite
aiofiles
uvloop
python-dotenv
git+https://github.com/FasterSpeeding/Sake